# kernel: Fix sleep timing accuracy