This PR corrects some mistake while comparing my reversing of kernel 13.x with our own.

``WaitAndCheckScheduledObjects`` timing accuracy was also improved.