@gdkchan 2021-11-09
Could potentially unblock #1798.