@riperiperi 2021-11-09
Greatly improves loading speeds in BOTW 1.0.0. Before they were very long and inconsistent, now they are short (3-5x faster) and consistent. This will likely also help ingame streaming, and I noticed a small performance boost ingame. (36-38 -> 40-42)

I took some rough measurements from booting the game to loading a save. "To Load Game" is the time until the save game starts loading (spamming a), "To 100% progress" is the time for the BOTW logo to fill entirely on the loading screen, and "To Ingame" is the time from that to the game beginning to fade in.

I noticed that with a hack to make the game run at 30 instead of 20, "to ingame" shortens to about 5.5 seconds. 

## PR

to load game: 15.42s
to 100% progress: 21.85s
to ingame: 8.74s

to load game: 14.78s
to 100% progress: 22.38s
to ingame: 8.89s

to load game: 14.55s
to 100% progress: 22.18s
to ingame: 8.89s

## Master

to load game: 15.00s
to 100% progress: 1:01.00s
to ingame: 9.58s

to load game: 15.34s
to 100% progress: 1:36.42s
to ingame: 15.27s

to load game: 14.37s
to 100% progress: 59.75s
to ingame: 7.96s