@Mou-Ikkai 2021-11-09
Hatsune Miku: Project DIVA Megamix seems to no longer crash for the first time on boot, forcing you to press stop emulation and rerun! 
Master: 10/10 times it froze on first run, but ran properly after "Stop Emulation" and relaunching the game.
PR: 0/10 times it froze on first run. This has fixed the first-time startup crash!