@jojorne 2021-11-10
Tested. Fixed boot on all my games.