@vickyorlo 2021-11-12
Incidentally, this fixes #2744, and all the other issues in Hyrule Warriors like the timer running slow, various mission progression softlocks/lagginess, and enemy spawning speed. However it feels like the game's performance has dropped - I'm getting framerates below 50 in busy scenes (which, considering it's a musou game, is almost always) when I used to see 60fps with drops to maybe 55 most of the time. It's hard to say, since with this fix the amount of enemies on screen has increased.