@d3xMachina 2021-11-19
Pokemon Brillant Diamond does not start with this PR. It stay blocked at a black screen even after waiting 5 minutes.
Also if I try to do "Stop emulation" and close Ryujinx while there is the black screen, it does not close Ryujinx unless I close the command line window.

On the latest Ryujinx version it start after like 1 min and the following messages are displayed before loading : 
> 00:00:21.405 |E| HLE.OsThread.9 Gpu WaitOnSyncpoint: Wait on syncpoint 1 for threshold 2 took more than 1000ms, resuming execution...
> 00:01:01.362 |E| HLE.OsThread.9 Gpu WaitOnSyncpoint: Wait on syncpoint 1 for threshold 6 took more than 1000ms, resuming execution...

Here are my logs (not much to see) : 
[Ryujinx_1.0.0+10c19b1.log](https://github.com/Ryujinx/Ryujinx/files/7572775/Ryujinx_1.0.0%2B10c19b1_2021-11-19_21-14-05.-.Copie.log)

The logs after 5 mins are after clicking stop emulation.