@d3xMachina 2021-11-20
I made a fix for the overflow on the timer.
Also I changed the wait so that the last millisecond we wait is done by spinning even when the wait is higher than 1ms for more precision.
No issue so far.

Here is the patch (rename the extension to .patch) :
[timer_patch.txt](https://github.com/Ryujinx/Ryujinx/files/7573759/timer_patch.txt)