@Thog 2021-11-23
> I made a fix for the overflow on the timer. Also I changed the wait so that the last millisecond we wait is done by spinning even when the wait is higher than 1ms for more precision. ScheduleFutureInvocation now only interrupt the wait if necessary. No issue so far.
> 
> Here is the patch (rename the extension to .patch) : [timer_patch.txt](https://github.com/Ryujinx/Ryujinx/files/7574436/timer_patch.txt)
> 
> Edit : I added a _waitEvent.Reset() in WaitAndScheduledObject in lock to avoid the case where _waitEvent.Set() is triggered again after _waitEvent.WaitOne() has been interrupted which would cause the next wait to be instantly canceled if higher than 1 ms.
> 
> I didn't completly explain why my patch fix the issue. The bug there was with the code is that _enforceWakeupFromSpinWait was only triggered if another timeout was of less than 1 ms. Instead it should have been canceled when a new scheduled invocation started sooner. With the overflow it would go spinning and only _enforceWakeupFromSpinWait was able to make it go out since the scheduled time was far in the future.

Fixed the issue on https://github.com/Ryujinx/Ryujinx/pull/2828/commits/1f175bf452075b46e6ef7068585afaae7db54228.
I fixed it by removing the unnecessary cast and by capping the upper timeout limit to 0x7FFFFFFF like it used to be. (see ConvertNanosecondsToMilliseconds behaviour)